Refactor: Organize check scripts into correct categories
I moved several check scripts to more appropriate directories:

- `netherRoofCheck.js` moved from `movement/` to `world/`
- `antiGMCCheck.js` moved from `world/` to `player/`
- `nameSpoofCheck.js` moved from `world/` to `player/`

I updated JSDoc `@file` tags in the moved files to reflect their new paths.
I adjusted export statements in `AntiCheatsBP/scripts/checks/index.js` to ensure all checks are correctly imported and exported from their new locations and categorized appropriately.
