+++ AntiCheatsBP/scripts/checks/index.js
@@ -11,7 +11,6 @@ export * from './movement/speedCheck.js';
 export * from './movement/noFallCheck.js';
 export { checkNoSlow } from './movement/noSlowCheck.js';
 export { checkInvalidSprint } from './movement/invalidSprintCheck.js';
-export * from './movement/netherRoofCheck.js';
 
 // Combat Checks
 export * from './combat/reachCheck.js';
@@ -27,15 +26,16 @@ export { checkTower, checkFlatRotationBuilding, checkDownwardScaffold, checkAirP
 export { checkFastUse } from './world/fastUseCheck.js';
 export { checkAutoTool } from './world/autoToolCheck.js';
 export { checkBreakUnbreakable, checkBreakSpeed } from './world/instaBreakCheck.js';
-export { checkNameSpoof } from './world/nameSpoofCheck.js';
-export { checkAntiGMC } from './world/antiGMCCheck.js';
 export { checkEntitySpam } from './world/entityChecks.js';
 export { checkPistonLag } from './world/pistonChecks.js';
+export * from './world/netherRoofCheck.js';
 
 // Player Behavior Checks
 export { checkSwitchAndUseInSameTick, checkInventoryMoveWhileActionLocked } from './player/inventoryModCheck.js';
 export { checkSelfHurt } from './player/selfHurtCheck.js';
 export { checkInvalidRenderDistance } from './player/clientInfoChecks.js';
+export { checkAntiGMC } from './player/antiGMCCheck.js';
+export { checkNameSpoof } from './player/nameSpoofCheck.js';
 
 // Chat Checks
 export { checkMessageRate } from './chat/messageRateCheck.js';
